832 Fix PyTorch 1.6 integration test error (#833)
* [DLMED] fix PyTorch 1.6 integration test error

* [DLMED] fix unit test

* [DLMED] update float64

* [DLMED] fix unit test

* [MONAI] python code formatting

* [DLMED] remove test branch

* [DLMED] make dtype config

* [DLMED] update test targets

* [DLMED] add default to float64

* [DLMED] add dtype to norm

* [DLMED] update dtype for all related modules

* [DLMED] update nifti writer

* [DLMED] restore norm

* [DLMED] update inference test

* [MONAI] python code formatting

* [DLMED] remove temp test branch

+++ tests/test_rotated.py
@@ -78,7 +78,7 @@ class TestRotated3D(NumpyImageTestCase3D):
         expected = scipy.ndimage.rotate(
             self.imt[0, 0], angle, (0, 2), not keep_size, order=_order, mode=_mode, prefilter=False
         )
-        np.testing.assert_allclose(expected, rotated["img"][0], atol=1e-3)
+        np.testing.assert_allclose(expected.astype(np.float32), rotated["img"][0], atol=1e-3)
 
         expected = scipy.ndimage.rotate(
             self.segn[0, 0], angle, (0, 2), not keep_size, order=0, mode=_mode, prefilter=False
